Features Added and Bug Fixes
-Added the ability to use argv when testing files.  To use this, the
first argument is the file to test, and an optional second argument is
used as an argument for the file.
-Added an option to type in a file name at the beginning of the program,
as well as arguments for the file.  The default test file takes input,
and is able to use this in place of typing input each time.
-Made the keyboard always visible so that it doesn't need to take time
being pulled away.
-Reduced the size of the bottom screen's text box so that text does not
display under the keyboard.
-Separated the top screen's debug information to 2 new files to help
with readability.
-Added a "stringInput" variable to the getC method.  This allows for an
entire array of characters to be stored here and read from instead of
requiring the keyboard to be used for input for this method.
-Updated test.asm to print out the correct number of lines for text
input (always printed 1 less before).